For indirect calls and stores, reorder ops if possible
These take different orders of arguments between AVM2 and Wasm,
so we have to reorder the arguments while evaluating them to
the stack.

A conservative check for side effects is done, since usually
there aren't any and you're just loading locals or doing a small
non-destructive op on the output from current transforms.

For indirect calls, temporary locals are used to store the
arguments to the call before evaluating the target.
Currently this will break if used with --trace, an exception
will be thrown.